requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
solders>=0.21.0
base58>=2.1.1
cryptography>=42.0.0
bcrypt>=4.1.0
stripe>=8.0.0
gunicorn>=21.2.0
gevent>=24.2.1
websocket-client>=1.7.0
psycopg2-binary>=2.9.9
kafka-python>=2.0.2
pandas>=2.0.0
ray>=2.10.0
